feat(lib): Move terms function to proper place
This 'view' was not really a view, because it did not produce results
based on the variable store. Instead it simply took parameters and
returned an object that was a valid terms query.

This is a good feature to have, but it should be internal to this
library only.
