+++ lib/leaf/terms.js
